Prepare to split the outcome for Kosovo
The current outcome template for opposite sex ceremonies in Kosovo
(outcome_opposite_sex_marriage_in_kosovo_when_residing_in_uk_or_kosovo)
contains two logical outcomes:

* One for residents of the UK
* One for residents of Kosovo

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.