Prepare for splitting the Kuwait outcome
The current outcome template (outcome_opposite_sex_marriage_in_kuwait)
contains three logical outcomes:

* One for residents of the UK
* One for residents of Kuwait
* One for residents of a third country

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.