Prepare to split marriage-abroad outcome for Gibraltar
The current outcome template
(outcome_same_sex_marriage_and_civil_partnership_not_possible) contains two
logical outcomes:

* One for ceremonies in Gibraltar
* One for countries where same sex ceremonies are not possible

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.