Prepare to split marriage-abroad outcome for Germany
The current outcome for opposite sex ceremonies in Germany
(outcome_opposite_sex_marriage_in_germany) contains two logical outcomes:

* One for residents of the UK
* One for residents outside the UK

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.