Update fees for Tunisia: Overseas Passports
Update the fees for Tunisia according to the specifications
document[1].

I have changed the application type for Tunisia as the new fees have
the same values as group 3. All failures for Tunisia have disappeared
so I trust that this is correct.

No failing tests for artefacts are expected at this point.