Refactor purpose of visit logic.
Sharing this logic with the channel island route

Here I have moved the the blocks and logic within the purpose of visit multiple choice question to the travel_response_next_route method (an instance method on CheckUkVisaFlow).

The reason for this is because as part of the new requirement for the Common travel areas.

A third new option has been introduced to clarify the transit rules that apply.

For users travelling to the Channel island or isle of man they will be presented with purpose of visit question without the transit option.

The logic here remains the same with purpose of visit (i.e Question 2).

Consequently, fixed failing tests